+++ nichtparasoup/config/__init__.py
@@ -1,15 +1,27 @@
-__all__ = ["get_config", "get_defaults", "dump_defaults", "get_imagecrawler", "parse_yaml_file"]
+__all__ = ["get_config", "get_defaults", "dump_defaults", "get_imagecrawler", "parse_yaml_file",
+           "ImageCrawlerSetupError"]
 
 from os.path import dirname, join as path_join, realpath
 from typing import Any, Dict, Optional
 
 from nichtparasoup.core.imagecrawler import BaseImageCrawler
 
-_schema_file = realpath(path_join(dirname(__file__), "schema.yaml"))
+_SCHEMA_FILE = realpath(path_join(dirname(__file__), "schema.yaml"))
 _schema = None  # type: Optional[Any]
 
-_defaults_file = realpath(path_join(dirname(__file__), "defaults.yaml"))
+_DEFAULTS_FILE = realpath(path_join(dirname(__file__), "defaults.yaml"))
 _defaults = None  # type: Optional[Dict[str, Any]]
+
+
+class ImageCrawlerSetupError(Exception):
+
+    def __init__(self, ic_name: str, ic_class: type, ic_config: Dict[Any, Any]) -> None:  # pragma: no cover
+        self._name = ic_name
+        self._class = ic_class
+        self._config = ic_config
+
+    def __str__(self) -> str:  # pragma: no cover
+        return 'Failed setup crawler {!r} of type {!r} with config {!r}'.format(self._name, self._class, self._config)
 
 
 def get_imagecrawler(config_crawler: Dict[str, Any]) -> BaseImageCrawler:
@@ -20,20 +32,16 @@ def get_imagecrawler(config_crawler: Dict[str, Any]) -> BaseImageCrawler:
         raise ValueError('unknown crawler name {!r}'.format(imagecrawler_name))
     imagecrawler_config = config_crawler['config']
     try:
-        imagecrawler_obj = imagecrawler_class(**imagecrawler_config)
+        return imagecrawler_class(**imagecrawler_config)
     except Exception as e:
-        raise Exception(
-            'failed setup crawler {!r} of type {!r} with config {!r}'
-            .format(imagecrawler_name, imagecrawler_class, imagecrawler_config)
-        ) from e
-    return imagecrawler_obj
+        raise ImageCrawlerSetupError(imagecrawler_name, imagecrawler_class, imagecrawler_config) from e
 
 
 def parse_yaml_file(file_path: str) -> Dict[str, Any]:
     import yamale  # type: ignore
     global _schema
     if not _schema:
-        _schema = yamale.make_schema(_schema_file, parser='ruamel')
+        _schema = yamale.make_schema(_SCHEMA_FILE, parser='ruamel')
     _data = yamale.make_data(file_path, parser='ruamel')
     config = yamale.validate(_schema, _data, strict=True)[0][0]  # type: Dict[str, Any]
     config.setdefault('logging', dict())
@@ -46,13 +54,13 @@ def parse_yaml_file(file_path: str) -> Dict[str, Any]:
 
 def dump_defaults(file_path: str) -> None:
     from shutil import copyfile
-    copyfile(_defaults_file, file_path)
+    copyfile(_DEFAULTS_FILE, file_path)
 
 
 def get_defaults() -> Dict[str, Any]:
     global _defaults
     if not _defaults:
-        _defaults = parse_yaml_file(_defaults_file)
+        _defaults = parse_yaml_file(_DEFAULTS_FILE)
     from copy import deepcopy
     return deepcopy(_defaults)