+++ tests/test_nichtparasoup/test_config/test_default_config_file.py
@@ -1,9 +1,0 @@
-
-from nichtparasoup.config import _defaults_file
-from nichtparasoup.testing.config import ConfigFilesTest
-
-
-class DefaultConfigFileTest(ConfigFilesTest):
-
-    def test_examples(self) -> None:
-        self.validate(_defaults_file)