+++ HISTORY.md
@@ -3,14 +3,21 @@
 
 ## Unreleased
 
+### Breaking changes
+
+* `nichtparasoup config --check`'s "duplicate image crawler" is no longer a Warning but an Error.
+* renamed `nichtparasoup.testing.config.ConfigFileTest` to `ConfigFileTest` - without an "s".
+
 ### Changes
 
+* `nichtparasoup config --check` now does a probe crawl.
 * class `ImageCrawlerInfo` lost support for positional arguments, supports keyword-arguments only - prepare future extensibility.
 * class `Image` lost support for positional arguments, supports keyword-arguments only - prepare future extensibility.
 
 ### Added
 
 * added image crawler for [pr0gramm](https://pr0gramm.com) - Read the [docs](./docs/imagecrawlers/pr0gramm.md).
+* added additional test function: `nichtparasoup.testing.config.ConfigFileTest.probe()`.
 
 
 ## 2.2.2